Make spikes one-way and reduce its size
- Remove corner spikes
- Change filled corners source id to 10

+++ src/map/spikes/spikes.gd
@@ -1,8 +1,10 @@
 class_name Spikes
 extends Area2D
 
+@onready var safe_area: Area2D = $SafeArea as Area2D
+
 func _on_body_entered(body: Node2D) -> void:
 	var player: Player = body as Player
 	
-	if player:
-		player.kill()
+	if player and not safe_area.overlaps_body(player):
+		player.kill()
